Feature/import rules

+++ src/webapp/pages/mappings/MappingsList.tsx
@@ -10,20 +10,21 @@ import {
     ConfirmationDialog,
     useSnackbar,
 } from "d2-ui-components";
+import AddIcon from "@material-ui/icons/Add";
 import Mapping from "../../models/Mapping";
 import i18n from "../../locales";
-import { useAppContext } from "../../contexts/app-context";
+import { useAppContext, useCompositionRoot } from "../../contexts/app-context";
 import { makeStyles } from "@material-ui/styles";
-import { Theme, createStyles, LinearProgress, Icon } from "@material-ui/core";
-import { withSnackbarOnError } from "../../utils/error";
+import { Theme, createStyles, LinearProgress, Icon, Box, Fab } from "@material-ui/core";
 import { useGoTo, GoTo, pageRoutes } from "../root/Root";
+import { DeleteMappingByIdsError } from "../../../domain/repositories/MappingRepository";
 
 type ContextualAction = "details" | "edit" | "delete";
 
 interface MappingsListProps {
     header?: string;
-    selectedMappings?: Mapping[];
-    onSelectionChange: (selectedMappings: Mapping[]) => void;
+    selectedMappings?: string[];
+    onSelectionChange: (selectedMappings: string[]) => void;
     onDeleteMappings: (deletedMappings: string[]) => void;
 }
 
@@ -106,9 +107,11 @@ const MappingsList: React.FC<MappingsListProps> = props => {
     const [isDeleting, setDeleting] = useState(false);
     const [objectsTableKey] = useState(() => new Date().getTime());
 
+    const mapping = useCompositionRoot().mapping();
+
     const selection = useMemo(() => {
         return rows
-            .filter(mapping => selectedMappings?.map(m => m.id).includes(mapping.id))
+            .filter(mapping => selectedMappings?.includes(mapping.id))
             .map(mapping => ({ id: mapping.id }));
     }, [rows, selectedMappings]);
 
@@ -132,29 +135,39 @@ const MappingsList: React.FC<MappingsListProps> = props => {
         setLoading(false);
     }
 
-    const deleteMappings = React.useCallback(() => {
+    const deleteMappings = async () => {
+        if (!mappingIdsToDelete) return;
+
         setDeleting(true);
-        withSnackbarOnError(
-            snackbar,
-            async () => {
-                await Mapping.delete(api, config, mappingIdsToDelete ?? []);
+
+        const handleFailure = (failure: DeleteMappingByIdsError): string => {
+            switch (failure.kind) {
+                case "UnexpectedError":
+                    return (
+                        i18n.t("An unexpected error has ocurred deleting mappings: ") +
+                        failure.error.message
+                    );
+            }
+        };
+
+        const results = await mapping.delete.execute(mappingIdsToDelete);
+
+        results.fold(
+            error => snackbar.error(handleFailure(error)),
+            () => {
                 onDeleteMappings(mappingIdsToDelete ?? []);
                 snackbar.success(
-                    i18n.t("{{n}} mappings deleted", {
-                        n: mappingIdsToDelete ? mappingIdsToDelete.length : 0,
+                    i18n.t("Successfully delete {{count}} mappings", {
+                        count: mappingIdsToDelete.length,
                     })
                 );
-            },
-            {
-                onFinally: () => {
-                    setDeleting(false);
-                    setMappingIdsToDelete(undefined);
-                    getMappings(sorting, { page: 1 });
-                },
             }
         );
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [api, config, snackbar, sorting, mappingIdsToDelete]);
+
+        setDeleting(false);
+        setMappingIdsToDelete(undefined);
+        getMappings(sorting, { page: 1 });
+    };
 
     const closeDeleteDialog = useCallback(() => {
         setMappingIdsToDelete(undefined);
@@ -163,7 +176,7 @@ const MappingsList: React.FC<MappingsListProps> = props => {
     const onTableChange = useCallback(
         (newSelectedMappingsIds: string[]) => {
             const newSelectedMappings = _.filter(rows, m => newSelectedMappingsIds.includes(m.id));
-            onSelectionChange(newSelectedMappings);
+            onSelectionChange(newSelectedMappings.map(mapping => mapping.id));
         },
         [rows, onSelectionChange]
     );
@@ -176,7 +189,7 @@ const MappingsList: React.FC<MappingsListProps> = props => {
                     onCancel={isDeleting ? undefined : closeDeleteDialog}
                     title={i18n.t("Delete mapping")}
                     description={i18n.t(
-                        "This operation will delete ({{n}}) mappings. This operation cannot be undone. Are you sure you want to proceed?",
+                        "This operation will delete ({{n}}) mappings and remove it as selected in related import rules. This operation cannot be undone. Are you sure you want to proceed?",
                         { n: mappingIdsToDelete.length }
                     )}
                     saveText={isDeleting ? i18n.t("Deleting...") : i18n.t("Proceed")}
@@ -187,23 +200,33 @@ const MappingsList: React.FC<MappingsListProps> = props => {
                 </ConfirmationDialog>
             )}
             {rows && (
-                <ObjectsTable<Mapping>
-                    key={objectsTableKey}
-                    selection={selection}
-                    searchBoxLabel={i18n.t("Search by name or code")}
-                    onChange={state => onTableChange(state.selection.map(m => m.id))}
-                    forceSelectionColumn={true}
-                    pagination={pagination}
-                    details={componentConfig.details}
-                    columns={componentConfig.columns}
-                    actions={componentConfig.actions}
-                    onActionButtonClick={() => goTo(pageRoutes.mappingsNew)}
-                    mouseActionsMapping={mouseActionsMapping}
-                    rows={rows}
-                    filterComponents={
-                        header && <div className={classes.tableHeader}>{header}:</div>
-                    }
-                />
+                <Box display="flex" flexDirection="column">
+                    {header && <div className={classes.tableHeader}>{header}:</div>}
+
+                    <ObjectsTable<Mapping>
+                        key={objectsTableKey}
+                        selection={selection}
+                        searchBoxLabel={i18n.t("Search by name or code")}
+                        onChange={state => onTableChange(state.selection.map(m => m.id))}
+                        forceSelectionColumn={true}
+                        pagination={pagination}
+                        details={componentConfig.details}
+                        columns={componentConfig.columns}
+                        actions={componentConfig.actions}
+                        mouseActionsMapping={mouseActionsMapping}
+                        rows={rows}
+                        filterComponents={
+                            <Fab
+                                className={classes.addButton}
+                                size="small"
+                                aria-label="add"
+                                onClick={() => goTo(pageRoutes.mappingsNew)}
+                            >
+                                <AddIcon />
+                            </Fab>
+                        }
+                    />
+                </Box>
             )}
         </div>
     );
@@ -213,7 +236,12 @@ const useStyles = makeStyles((theme: Theme) =>
     createStyles({
         tableHeader: {
             ...theme.typography.button,
-            padding: theme.spacing(1),
+            marginTop: theme.spacing(5),
+            marginLeft: theme.spacing(1),
+        },
+        addButton: {
+            marginLeft: theme.spacing(1),
+            boxShadow: "none",
         },
     })
 );
